requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
python-dotenv==1.2.1

httpx==0.27.2
groq==1.0.0

python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator

transformers==4.49.0
torch>=2.1.0
soundfile==0.13.1
numpy==2.2.3
